Fix nested exception unwind
When we take an exception inside of an except: block we currently don't
unwind the current exception.  This results in us leaving the except block
and believing there is still an active exception in flight.

This fixes that, which is the equivalent of adding ceval.c's:

            if (b->b_type == EXCEPT_HANDLER) {
                   UNWIND_EXCEPT_HANDLER(b);
                   continue;
	    }

in fast_block_end