More exception handling fixes.
Fix for test_contextlib, where we weren't properly unwinding
exceptions in the nested exception case.  This resulted in an
infinite loop while trying to restore the current exception
as we had a loop of exception causes.

Now we unwind exceptions through all of the handlers which are
impacted on the current stack.

Also includes some fixes for code generated under a double
try/finally blocks.  When we're in the finally block we have an
extra value on the stack.  Our code gen generally works fine with
a single finally, but when we have 2 of them we aren't accounting
for the value that's just hanging out on the stack.  Now we use
BranchLeave to clear the stack before branching out of the finally.