Fix #120
The JIT is emitting lots of temporary spills when we do a branch
with a value on the stack.  This leads to a huge frame size.  Most
of these branches are happening on our branches for error checks.

This stores the value in a temp and reloads it instead of branching
with the value on the stack.  This greatly reduces the number of
these, but once we get something like:

a() + b()

Where we do an error check on the result of b() we still have the
result of a() on the stack, which causes us to still spill.  This
at least gets us passing on this test, but we'll need to see if
we can be more efficient in the future, or if we just need to store
all of the values around an error check.