Force the unary inverts operand to escape
The current code generator is not wired up to handle unboxed
values for invert.  In the long term, proper support for this
should be added, but in the near term an escape should be
forced so that the code is correct.

This was discovered by doing a test suite run and forcing JIT
compilation for every test.