Add JIT specialization threshold option
The JIT test suite is currently *not* exercising the JIT engine
because the specialization hit count threshold is never exceeded.
This patch adds an option to `PyjionJittedCode` that allows the
threshold to be specified instead of hard coding it in `Jit_EvalTrace`.
This way the test code can set it to 0 thereby forcing the code to
always be JITed.