Improve ROT_THREE support
This patch changes the implementation of ROT_THREE to only
escape its operands when the types are not all ints or
all floats.  Furthermore, the ROT_THREE code generator has
been enhanced to use the appropriate type for defining
locals.

This is the companion patch to 539866f0ec5e3595bfd5e8c9117084f2517dbc58
and is the last patch to properly resolve issue #88.